Splice out DAW component
